jBack/web/serializer/room_serializer.py
from rest_framework import serializers
from jBack.web.model import Room

class RoomSerializer(serializers.ModelSerializer):
    class Meta:
        model = Room
        fields = ['id', 'name', 'max_occupancy', 'start_time', 'end_time', 'description']
        read_only_fields = ('id',)

    def create(self, validated_data):
        return Room.objects.create(**validated_data)

    def update(self, instance, validated_data):
        instance.name = validated_data.get('name', instance.name)
        instance.description = validated_data.get('description', instance.description)
        instance.max_occupancy = validated_data.get('max_occupancy', instance.max_occupancy)
        instance.start_time = validated_data.get('start_time', instance.start_time)
        instance.end_time = validated_data.get('end_time', instance.end_time)
        instance.save()
        return instance

