requirements.txt
Django
psycopg2-binary>=2.8
djangorestframework
python-dateutil==2.8.2
django-cors-headers==4.3.1
djangorestframework-api-key